% Auto-generated: do not edit by hand
\name{dbcModalFooter}

\alias{dbcModalFooter}

\title{ModalFooter component}

\description{
Add a footer to any modal.
}

\usage{
dbcModalFooter(children=NULL, id=NULL, style=NULL, className=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{style}{Named list. Defines CSS styles which will override styles previously set.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{tag}{Character. HTML tag to use for the ModalFooter, default: div}
}
