% Auto-generated: do not edit by hand
\name{dbcModalHeader}

\alias{dbcModalHeader}

\title{ModalHeader component}

\description{
Add a header to any modal.
}

\usage{
dbcModalHeader(children=NULL, id=NULL, className=NULL, style=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{style}{Named list. Defines CSS styles which will override styles previously set.}

\item{tag}{Character. HTML tag to use for the ModalHeader, default: div}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

